requirements.txt
Flask
supabase
scikit-learn
pandas
numpy
psycopg2-binary
python-dotenv
